PostProcessEffect: refactoring
